updated name + redirect fix + UI changes
