+++ frontend/src/components/admin/menu/index.js
@@ -3,4 +3,4 @@ export { default as BillingMenuManagement } from "./BillingMenuManagement";
 export { default as NonConformityMenuManagement } from "./NonConformityMenuManagement";
 export { default as PatientMenuManagement } from "./PatientMenuManagement";
 export { default as StudyMenuManagement } from "./StudyMenuManagement";
-export { default as SampleEntryConfigMenu } from "./SampleEntryConfigMenu";
+export { default as OrderEntryConfigMenu } from "./OrderEntryConfigMenu";
